+++ src/libraries/Pdf.php
@@ -3,19 +3,28 @@ namespace amici\SuperPdf\libraries;
 
 use Craft;
 use amici\SuperPdf\SuperPdf;
-use Dompdf\Dompdf;
+use amici\SuperPdf\models\Asset as PdfAsset;
 
+use craft\helpers\DateTimeHelper;
 use craft\helpers\FileHelper;
 use craft\helpers\UrlHelper;
+use craft\helpers\Assets;
+use craft\elements\Asset;
+use craft\elements\User;
+
 use yii\base\Exception;
+use Dompdf\Dompdf;
 
 class Pdf
 {
+	public $asset;
 
 	public $html = "";
 	public $settings = [];
 	public $dompdf;
 	public $devMode;
+	public $folder;
+	public $path;
 
 	function __construct()
 	{
@@ -23,31 +32,39 @@ class Pdf
 		$this->settings = SuperPdf::$plugin->getSettings()->getAttributes();
 	}
 
-	function html($html = "", $settings = [])
+	public function html($html = "", $settings = [])
 	{
 
 		foreach ($settings as $key => $value)
 		{
 			$this->settings[$key] = $value;
 		}
+
+		$this->fetchFile();
 
 		$this->html = $html;
 		return $this->_generate();
 
 	}
 
-	function template($template = "", $settings = [], $vars = [])
+	public function template($template = "", $settings = [], $vars = [])
 	{
-
 		foreach ($settings as $key => $value)
 		{
 			$this->settings[$key] = $value;
 		}
 
-		// Craft::$app->getView()->setTemplatesPath(Craft::$app->getPath()->getSiteTemplatesPath());
-		$this->html = Craft::$app->getView()->renderTemplate($template, $vars);
-		return $this->_generate();
+		$this->fetchFile();
 
+		if(
+			! $this->asset ||
+			$this->settings['resaveBehaviour'] != 'ignore' ||
+			! in_array($this->settings['type'], ['url', 'object'])
+		) {
+			$this->html = Craft::$app->getView()->renderTemplate($template, $vars);
+		}
+
+		return $this->_generate();
 	}
 
 	private function _generate()
@@ -55,6 +72,20 @@ class Pdf
 
 		try
 		{
+			if($this->asset && $this->settings['resaveBehaviour'] == 'ignore')
+			{
+				// Type URL is deprecated
+				if(in_array($this->settings['type'], ['url', 'object']))
+				{
+					return $this->asset;
+				}
+
+				if($this->settings['volume'] == "storage") {
+					return $this->renderFileFromStorage($this->settings['filename'] . '.pdf');
+				}
+
+				return $this->renderFileFromAssetsElement();
+			}
 
 			$this->dompdf = new Dompdf($this->settings);
 			$this->dompdf->setPaper($this->settings['defaultPaperSize'], $this->settings['defaultPaperOrientation']);
@@ -69,7 +100,6 @@ class Pdf
 
 			if($this->settings['encrypt'])
 			{
-
 				$allow = [];
 				if($this->settings['print']) 	$allow[] = 'print';
 				if($this->settings['modify']) 	$allow[] = 'modify';
@@ -77,7 +107,6 @@ class Pdf
 				if($this->settings['add']) 		$allow[] = 'add';
 
 				$this->dompdf->getCanvas()->get_cpdf()->setEncryption($this->settings['password'], $this->settings['adminPassword'], $allow);
-
 			}
 
 			if($this->settings['type'] == 'url')
@@ -85,12 +114,15 @@ class Pdf
 				return $this->url();
 			}
 
-			return $this->render();
+			if($this->settings['type'] == 'object')
+			{
+				return $this->object();
+			}
 
+			return $this->render();
 		}
 		catch (Exception $e)
 		{
-
 			$error = $e->getMessage();
 			SuperPdf::error($error);
 
@@ -102,17 +134,45 @@ class Pdf
 			{
 				return false;
 			}
-
 		}
 
 	}
 
-	function render()
+	public function renderFileFromStorage($filename = "")
 	{
+		if($filename == "")
+		{
+			$error = SuperPdf::t("Filename field cannot be empty");
+			SuperPdf::error($error);
+			throw new Exception($error);
+		}
 
+		$this->path = $this->getStoragePath() . '/' . $filename;
+		if(! file_exists($this->path))
+		{
+			$error = SuperPdf::t("PDF File does not exists.");
+			SuperPdf::error($error);
+			throw new Exception($error);
+		}
+
+		$response = Craft::$app->getResponse();
+		return $response->sendFile($this->path, null, ['inline' => true]);
+	}
+
+	public function renderFileFromAssetsElement()
+	{
+		return Craft::$app->getResponse()
+	        ->sendStreamAsFile($this->asset->asset->getStream(), $this->asset->asset->filename, [
+	            'fileSize' => $this->asset->asset->size,
+	            'mimeType' => $this->asset->asset->getMimeType(),
+	            'inline' => true
+	        ]);
+	}
+
+	private function render()
+	{
 		try
 		{
-
 			$options = array(
 				'Attachment' => $this->settings['forceDownload'],
 				'compress' 	 => $this->settings['compress']
@@ -121,11 +181,9 @@ class Pdf
 			ob_end_clean();
 			$this->dompdf->stream($this->settings['filename'], $options);
 			exit();
-
 		}
 		catch (Exception $e)
 		{
-
 			$error = $e->getMessage();
 			SuperPdf::error($error);
 
@@ -137,24 +195,71 @@ class Pdf
 			{
 				return false;
 			}
-
 		}
-
 	}
 
-	function url()
+	private function url()
 	{
+		Craft::$app->getDeprecator()->log(__METHOD__, "PDF Settings `type: url` setting has been deprecated. Use `type: object` instead.");
+		return $this->object();
+	}
 
+	private function object()
+	{
 		try
 		{
-			$filename = $this->settings['filename'] . '_' . rand(9999, 99999) . '.pdf';
-			$path = SuperPdf::$plugin->general->getStoragePath();
-			$path = $path . '/' . $filename;
-			FileHelper::writeToFile($path, $this->dompdf->output());
+			if($this->asset != '' && $this->settings['resaveBehaviour'] == "duplicate") {
+				$this->settings['filename'] = $this->settings['filename'] . '_' . rand(9999, 99999);
+			}
 
-			$pdfUrl = rtrim(UrlHelper::siteUrl(), '/') . '/super-pdf' . '/' . $filename;
-			// return $path . "<br>" . $pdfUrl;
-			return $pdfUrl;
+			$filename = $this->settings['filename'] . '.pdf';
+
+			$this->path = $this->getStoragePath(). '/' . $filename;
+			FileHelper::writeToFile($this->path, $this->dompdf->output());
+			if($this->settings['volume'] == "storage")
+			{
+				return $this->createAssetObject();
+				// return rtrim(UrlHelper::siteUrl(), '/') . '/super-pdf' . '/' . $filename;
+			}
+
+		    $assetsService = Craft::$app->getAssets();
+			$folder = $this->getFolder();
+			$asset = Asset::find()->filename($filename)->volumeId($folder->volumeId)->folderId($folder->id)->one();
+			if($asset)
+			{
+				$filename = Assets::prepareAssetName($filename);
+				$assetsService->replaceAssetFile($asset, $this->path, $filename);
+
+				return $this->createAssetObject();
+			}
+			else
+			{
+
+				$asset = new Asset();
+				$asset->tempFilePath = $this->path;
+				$asset->filename = $filename;
+				// $asset->title = AssetsHelper::filename2Title(pathinfo($filename, PATHINFO_FILENAME));
+				// $asset->newFolderId = $folder->id;
+				$asset->newLocation = "{folder:{$folder->id}}{$filename}";
+
+				// $asset->uploaderId = "1";
+				$asset->folderId = $folder->id;
+				$asset->volumeId = $folder->volumeId;
+				$asset->avoidFilenameConflicts = true;
+				$asset->setScenario(Asset::SCENARIO_CREATE);
+				$result = Craft::$app->getElements()->saveElement($asset);
+
+				if(! $result) {
+					if(! empty($asset->getErrors())) {
+						throw new Exception($this->getFirstError($asset->getErrors()), 1);
+					}
+
+					throw new Exception("We are not able to save the file at the moment. Make sure you have write permission to the folder", 1);
+				}
+
+				return $this->createAssetObject($asset);
+			}
+
 		}
 		catch (Exception $e)
 		{
@@ -175,4 +280,114 @@ class Pdf
 
 	}
 
+	private function fetchFile()
+	{
+		$filename = $this->settings['filename'] . '.pdf';
+
+		if($this->settings['volume'] == 'storage')
+		{
+			$this->path = $this->getStoragePath() . '/' . $filename;
+
+			if(file_exists($this->path))
+			{
+				return $this->createAssetObject();
+			}
+
+			return null;
+		}
+
+		$folder = $this->getFolder();
+		$asset = Asset::find()->filename($filename)->volumeId($folder->volumeId)->folderId($folder->id)->one();
+		if($asset)
+		{
+			return $this->createAssetObject($asset);
+		}
+
+		return null;
+	}
+
+	private function createAssetObject($asset = null)
+	{
+		$this->asset = new PdfAsset();
+
+		if($asset)
+		{
+			$this->asset->filename = $asset->filename;
+			$this->asset->kind = $asset->kind;
+			$this->asset->size = $asset->size;
+			$this->asset->dateModified = $asset->dateModified;
+			$this->asset->path = $asset->getPath();
+			$this->asset->url = $asset->getUrl();
+			$this->asset->asset = $asset;
+
+			return $this->asset;
+		}
+		else
+		{
+			$lastUpdated = @filemtime($this->path);
+			$this->asset = new PdfAsset();
+
+			$this->asset->filename = $this->settings['filename'] . '.pdf';
+			$this->asset->kind = "pdf";
+			$this->asset->size = @filesize($this->path);
+			$this->asset->dateModified = ($lastUpdated ? DateTimeHelper::toDateTime($lastUpdated) : null);
+			$this->asset->path = $this->path;
+			$this->asset->url = rtrim(UrlHelper::siteUrl(), '/') . '/super-pdf' . '/' . $this->asset->filename;
+			$this->asset->asset = null;
+
+			return $this->asset;
+		}
+
+		return null;
+	}
+
+	private function getStoragePath()
+	{
+		$path = SuperPdf::$plugin->general->getStoragePath();
+		return rtrim($path, '/') . ($this->settings['folder'] != '' ? '/' . $this->settings['folder'] : '');
+	}
+
+	private function getFolder()
+	{
+		if($this->folder) {
+			return $this->folder;
+		}
+
+		$volume = Craft::$app->getVolumes()->getVolumeByHandle($this->settings['volume']);
+
+		if(! $volume) {
+			throw new Exception("Volume \"{$this->settings['volume']}\" not found or no longer exists.", 1);
+		}
+
+		$assetsService = Craft::$app->getAssets();
+		if($this->settings['folder'] == '')
+		{
+			$this->folder = $assetsService->getRootFolderByVolumeId($volume->id);
+			return $this->folder;
+		}
+
+		$this->folder = $assetsService->findFolder([
+		    'volumeId'  => $volume->id,
+		    'path' 		=> $this->settings['folder'] . '/'
+		]);
+
+		if (! $this->folder) {
+		    $folderId = $assetsService->ensureFolderByFullPathAndVolume($this->settings['folder'], $volume);
+			$this->folder = $assetsService->findFolder(['id' => $folderId]);
+		}
+
+		return $this->folder;
+	}
+
+	private function getFirstError($errors)
+	{
+		foreach ($errors as $key => $value)
+		{
+			return $key . ": " . $value[0];
+
+			break;
+		}
+
+		return null;
+	}
 }