+++ src/models/Settings.php
@@ -47,4 +47,48 @@ class Settings extends Model
     public $type = 'render'; // url
     public $streamContext = [];
 
+    // General Settings
+    public $pluginName = "Super PDF";
+    public $hasCpSection = false;
+    public $volume = "storage";
+    public $folder = "";
+    public $resaveBehaviour = "duplicate";
+
+    public function getSettingsNavItems(): array
+    {
+        $ret = [];
+        if(Craft::$app->getConfig()->getGeneral()->allowAdminChanges) {
+            $ret +=  [
+                'local' => [
+                    'label'     => Craft::t('super-pdf', 'General Settings'),
+                    'url'       => 'super-pdf',
+                    'action'    => 'super-pdf/settings/save-general-settings',
+                    'redirect'  => 'super-pdf/settings',
+                    'selected'  => 'local',
+                    'template'  => 'super-pdf/_templates/general'
+                ]
+            ];
+        }
+
+        /*$ret +=  [
+            'truncate' => [
+                'label'     => Craft::t('super-pdf', 'Truncate'),
+                'url'       => 'super-pdf/truncate',
+                'action'    => 'super-pdf/settings/truncate-queue',
+                'redirect'  => 'super-pdf/truncate',
+                'selected'  => 'truncate',
+                'template'  => 'super-pdf/_templates/truncate'
+            ],
+        ];*/
+
+        return $ret;
+
+    }
+
+    public function rules(): array
+    {
+        return [
+            [['pluginName', 'volume', 'resaveBehaviour'], 'required'],
+        ];
+    }
 }