+++ src/services/App.php
@@ -1,0 +1,14 @@
+<?php
+namespace amici\SuperPdf\services;
+
+use craft\base\Component;
+
+class App extends Component
+{
+	public $settings;
+
+	public function init(): void
+    {
+        $this->settings = new Settings();
+    }
+}