+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## 1.0.8 - 2022-09-23
+> {warning} For "url" type, PDF code will return object instead of URL. Code will not break but advisable to use {{ object.getUrl() }} instead {{ object }}.
+> {warning} New Settings section added to give user more control over PDF behaviour. Make sure to review and change settings according to your needs.
+- Adding Settings to give user more control over the plugin.
+- Introducing filesystems so user can store pdfs outside of storage folder.
+- Depending on the settings, We can now set wether we want to regenerate file on each page load, override it or ignore new file creation.
+- Type "url" is deprecated. Use "object" instead.
+- Object can be use with variables or methods to get the URL or other meta data from the file. For example:
+```
+{% set object = craft.superpdf.template("pdf-template", settings, vars) %}
+{{ object.url }}
+{{ object.getUrl() }}
+
+{{ object.path }}
+{{ object.getPath() }}
+
+{{ object.filename }}
+{{ object.getFilename() }}
+
+{{ object.kind }}
+{{ object.size }}
+{{ object.dateModified|date("m/d/Y H:i:s") }}
+
+// Returns craft assets element. Only if PDF is stored in craft assets volumes instead of storage folder.
+{{ object.asset }}
+{{ object.getAsset() }}
+```
+
 ## 1.0.7 - 2022-03-09
 - Added new setting `streamContext` where user can pass any HTTP Context. For example:
 ```